refactor: #397 TReservationInfoController を責務ごとに分割する
2043行のモノリシックコントローラーを6つの専用クラスに分割した。

分割後の構成:
- ReservationBaseController: 共有テーブル・サービス・authorizeReservation を提供する抽象基底
- ReservationToggleController: toggle() のみ（Kid UI 向け食事フラグ ON/OFF）
- ReservationBulkController: 週次一括登録・直前一括編集（BulkActionsTrait を吸収）
- ReservationCopyController: 週・月コピーAPI（CopyActionsTrait を吸収）
- ReservationReportController: 食数取得・エクスポート・欠食/実食報告（ReportActionsTrait を吸収）
- ReservationActualMealController: 実食確認管理・食数グリッド
- TReservationInfoController: カレンダー・add・changeEdit のコアCRUD（2043→904行）

routes.php で URL を保持したまま各コントローラーへ振り分けを更新。
3つの Trait ファイルは各コントローラーに直接組み込み削除。
